add: 게시물 삭제 로직 작성 완료

+++ src/boards/application/service/boards.service.ts
@@ -5,15 +5,18 @@ import { CreateBoardResponse } from '../dto/response/create-board.response';
 import { ReadBoardResponse } from '../dto/response/read-board.response';
 import { PrismaService } from '../../../prisma/prisma.service';
 import { DeleteBoardResponse } from '../dto/response/delete-board.response';
+import { BoardValidator } from '../validator/board-validator';
 
 @Injectable()
 export class BoardsService {
   constructor(
     @Inject('BoardRepository')
     private readonly boardRepository: BoardRepository,
+    private readonly boardValidator: BoardValidator,
     private readonly prismaService: PrismaService,
   ) {}
 
+  // 게시글 생성
   async create(
     createRequest: CreateBoardRequest,
   ): Promise<CreateBoardResponse> {
@@ -22,13 +25,20 @@ export class BoardsService {
     );
   }
 
+  // 게시글 상세 조회
   async getDetailBoard(boardId: number): Promise<ReadBoardResponse | null> {
     return this.prismaService.$transaction(async () =>
       this.getDetailBoardTransaction(boardId),
     );
   }
 
-  // 게시글 생성
+  // 게시글 삭제
+  async deleteBoard(boardId: number): Promise<DeleteBoardResponse> {
+    return this.prismaService.$transaction(async () =>
+      this.deleteBoardTransaction(boardId),
+    );
+  }
+
   private async createTransaction(
     createRequest: CreateBoardRequest,
   ): Promise<CreateBoardResponse> {
@@ -38,16 +48,21 @@ export class BoardsService {
     return CreateBoardResponse.fromEntity(createdBoard);
   }
 
-  // 게시글 조회
   private async getDetailBoardTransaction(
     boardId: number,
   ): Promise<ReadBoardResponse | null> {
-    const foundBoard = await this.boardRepository.findById(boardId);
+    const foundBoard = await this.boardRepository.findAliveBoardById(boardId);
     return foundBoard ? ReadBoardResponse.fromEntities(foundBoard) : null;
   }
 
-  // 게시글 삭제
-  async unregisterBoard(boardId: number) {
+  async deleteBoardTransaction(boardId: number) {
+    // 우선 삭제 가능한 상태인지 검증한다
+    const validationResult = await this.boardValidator.isDeletable(boardId);
+
+    if (!validationResult.success) {
+      throw validationResult.exception;
+    }
+
     const deleteBoard = await this.boardRepository.deleteById(boardId);
     return DeleteBoardResponse.fromEntities(deleteBoard);
   }
